Replaced all from integer to long

+++ src/main/java/executor/MyCallable.java
@@ -3,15 +3,15 @@ package executor;
 import java.util.List;
 import java.util.concurrent.Callable;
 
-public class MyCallable implements Callable<Integer> {
-    private List<Integer> numbers;
+public class MyCallable implements Callable<Long> {
+    private List<Long> numbers;
 
-    public MyCallable(List<Integer> numbers) {
+    public MyCallable(List<Long> numbers) {
         this.numbers = numbers;
     }
 
     @Override
-    public Integer call() throws Exception {
-        return numbers.stream().mapToInt(Integer::valueOf).sum();
+    public Long call() throws Exception {
+        return numbers.stream().mapToLong(Long::valueOf).sum();
     }
 }
